requirements.txt
numpy==2.3.3
pandas==2.3.3
bottleneck==1.6.0
numexpr==2.13.1
# Indicator libraries
ft-pandas-ta==0.3.16
ta-lib==0.6.7
technical==1.5.3

ccxt==4.5.10
cryptography==46.0.2
aiohttp==3.13.0
SQLAlchemy==2.0.44
python-telegram-bot==22.5
# can't be hard-pinned due to telegram-bot pinning httpx with ~
httpx>=0.24.1
humanize==4.13.0
cachetools==6.2.1
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
jsonschema==4.25.1
tabulate==0.9.0
pycoingecko==3.2.0
jinja2==3.1.6
joblib==1.5.2
rich==14.2.0
pyarrow==21.0.0; platform_machine != 'armv7l' and platform_machine != "aarch64"
# TODO: downgrade for aarch64 until https://github.com/apache/arrow/issues/47229 is resolved
pyarrow==20.0.0;  platform_machine == "aarch64"

# Load ticker files 30% faster
python-rapidjson==1.21
# Properly format api responses
orjson==3.11.3

# Notify systemd
sdnotify==0.3.2

# API Server
fastapi==0.119.0
pydantic==2.12.0
uvicorn==0.37.0
pyjwt==2.10.1
aiofiles==25.1.0
psutil==7.1.0

# Building config files interactively
questionary==2.1.1
prompt-toolkit==3.0.52
# Extensions to datetime library
python-dateutil==2.9.0.post0
pytz==2025.2

#Futures
schedule==1.2.2

#WS Messages
websockets==15.0.1
janus==2.0.0

ast-comments==1.2.3
packaging==25.0
PyYAML==6.0.2
ta==0.10.2
pytorch-forecasting==1.5.0
stable-baselines3==2.7.0
optuna==3.4.0
wandb==0.16.1
